refactoring

+++ d_server/inc/bme280_i2c.h
@@ -1,1 +1,1 @@
-int initialize_bme(char * i2c_loc, float * T, float * H);
+int initialize_bme(char * i2c_loc, double * T, double * H);
