+++ d_server/inc/client.h
@@ -1,4 +1,4 @@
 int init_client();
-int message(float * H, float * T, int lamp[], int ac[], int sp[], int so[]);
+int message(double * H, double * T, int lamp[], int ac[], int sp[], int so[]);
 int sensor_message(char sensor_data[]);
 void close_socket();
